requirements.txt
python==3.12.8
requests==2.31.0
streamlit==1.24.1
pandas==2.2.4   
python-dotenv==1.0.1